Separate TorchScript utils into a module
