+++ ftorch_utils/pt2ts.py
@@ -1,70 +1,14 @@
 #!/bin/env python
 """Convert a PyTorch model file to TorchScript format."""
 
+import argparse
+import os
+
 import torch
 
-
-def script_to_torchscript(model: torch.nn.Module, filename: str) -> None:
-    """
-    Save PyTorch model to TorchScript using scripting.
-
-    Parameters
-    ----------
-    model : torch.NN.Module
-        a PyTorch model
-    filename : str
-        name of file to save to
-    """
-    # FIXME: Adopt torch.jit.optimize_for_inference() once
-    # https://github.com/pytorch/pytorch/issues/81085 is resolved
-    scripted_model = torch.jit.script(model)
-    # print(scripted_model.code)
-    scripted_model.save(filename)
-
-
-def trace_to_torchscript(
-    model: torch.nn.Module,
-    input_tensor: torch.Tensor,
-    filename: str,
-) -> None:
-    """
-    Save PyTorch model to TorchScript using tracing.
-
-    Parameters
-    ----------
-    model : torch.NN.Module
-        a PyTorch model
-    input_tensor : torch.Tensor
-        appropriate size Tensor to act as input to model
-    filename : str
-        name of file to save to
-    """
-    # FIXME: Adopt torch.jit.optimize_for_inference() once
-    # https://github.com/pytorch/pytorch/issues/81085 is resolved
-    traced_model = torch.jit.trace(model, input_tensor)
-    # traced_model.save(filename)
-    frozen_model = torch.jit.freeze(traced_model)
-    ## print(frozen_model.graph)
-    ## print(frozen_model.code)
-    frozen_model.save(filename)
-
-
-def load_torchscript(filename: str) -> torch.nn.Module:
-    """
-    Load a TorchScript model from file.
-
-    Parameters
-    ----------
-    filename : str
-        name of file containing TorchScript model
-    """
-    return torch.jit.load(filename)
-
+from ftorch_utils.torchscript import script_to_torchscript, trace_to_torchscript
 
 if __name__ == "__main__":
-    import argparse
-    import os
-
     # Parse user input
     parser = argparse.ArgumentParser(
         prog="pt2ts.py",
